move process to compartment directory
