+++ vivarium/composites/glc_lct_shifter.py
@@ -139,7 +139,7 @@ def plot_diauxic_shift(timeseries, settings={}, out_dir='out'):
 
 
 if __name__ == '__main__':
-    from vivarium.actor.process import load_compartment
+    from vivarium.compartment.process import load_compartment
     from vivarium.compartment.composition import simulate_with_environment, convert_to_timeseries, plot_simulation_output
 
     out_dir = os.path.join('out', 'tests', 'glc_lct_shifter')
