+++ vivarium/composites/growth_division.py
@@ -2,7 +2,7 @@ from __future__ import absolute_import, division, print_function
 
 import os
 
-from vivarium.actor.process import initialize_state
+from vivarium.compartment.process import initialize_state
 from vivarium.compartment.composition import get_derivers, get_schema
 
 # processes
@@ -73,7 +73,7 @@ def compose_growth_division(config):
 
 
 if __name__ == '__main__':
-    from vivarium.actor.process import load_compartment
+    from vivarium.compartment.process import load_compartment
     from vivarium.compartment.composition import simulate_with_environment, convert_to_timeseries, plot_simulation_output
 
     out_dir = os.path.join('out', 'tests', 'growth_division_composite')
