+++ vivarium/composites/variable_flagella.py
@@ -4,7 +4,7 @@ import copy
 import os
 import random
 
-from vivarium.actor.process import initialize_state
+from vivarium.compartment.process import initialize_state
 from vivarium.compartment.composition import get_derivers, get_schema
 
 # processes
@@ -91,7 +91,7 @@ def compose_variable_flagella(config):
 
 
 if __name__ == '__main__':
-    from vivarium.actor.process import load_compartment
+    from vivarium.compartment.process import load_compartment
     from vivarium.compartment.composition import simulate_with_environment, convert_to_timeseries, plot_simulation_output
 
     out_dir = os.path.join('out', 'tests', 'variable_flagella_composite')
