+++ vivarium/processes/Endres2006_chemoreceptor.py
@@ -3,7 +3,7 @@ from __future__ import absolute_import, division, print_function
 import os
 import math
 
-from vivarium.actor.process import Process
+from vivarium.compartment.process import Process
 from vivarium.utils.units import units
