+++ vivarium/processes/Mears2014_flagella_activity.py
@@ -11,7 +11,7 @@ import matplotlib.pyplot as plt
 from matplotlib import colors
 from matplotlib.patches import Patch
 
-from vivarium.actor.process import Process
+from vivarium.compartment.process import Process
 from vivarium.compartment.composition import simulate_process_with_environment, convert_to_timeseries
 
 DEFAULT_N_FLAGELLA = 5
