+++ vivarium/processes/convenience_kinetics.py
@@ -4,7 +4,7 @@ import os
 
 from scipy import constants
 
-from vivarium.actor.process import Process
+from vivarium.compartment.process import Process
 from vivarium.compartment.composition import simulate_process_with_environment, convert_to_timeseries, plot_simulation_output
 from vivarium.utils.kinetic_rate_laws import KineticFluxModel
 from vivarium.utils.dict_utils import tuplify_role_dicts
