+++ vivarium/processes/growth.py
@@ -2,7 +2,7 @@ from __future__ import absolute_import, division, print_function
 
 import numpy as np
 
-from vivarium.actor.process import Process
+from vivarium.compartment.process import Process
 
 
 class Growth(Process):
