+++ vivarium/processes/transport_lookup.py
@@ -1,10 +1,9 @@
 from __future__ import absolute_import, division, print_function
 
 import os
-import csv
 from scipy import constants
 
-from vivarium.actor.process import Process
+from vivarium.compartment.process import Process
 from vivarium.environment.make_media import Media
 from vivarium.environment.look_up import LookUp
 from vivarium.utils.rate_law_utilities import load_reactions
